langchain/history_aware/requirements.txt
gradio
python-dotenv
langchain
langchain_community
PyPDF2
langchain_openai

faiss-cpu